Extract BottomModal animation logic & state in a hook
